refactor transactions saga
